New Version 0.2.0

+++ app/main.py
@@ -1,33 +1,39 @@
+from asyncio import AbstractEventLoop, get_running_loop
 from collections.abc import AsyncIterator, Awaitable, Callable
 from contextlib import asynccontextmanager
 
+from aio_pika import connect_robust
 from fastapi import FastAPI
 from starlette.requests import Request
 from starlette.responses import Response
 
 from app.common.config import (
-    DATABASE_RESET,
-    DB_URL,
+    DATABASE_MIGRATED,
+    MQ_URL,
     PRODUCTION_MODE,
     Base,
     engine,
+    pochta_producer,
     sessionmaker,
 )
 from app.common.sqla import session_context
-from app.routes import proxy_rst, reglog_rst, sessions_rst, users_mub, users_rst
+from app.routes import current_user_rst, proxy_rst, reglog_rst, sessions_rst, users_mub
 from app.utils.cors import CorrectCORSMiddleware
 
 
 @asynccontextmanager
 async def lifespan(_: FastAPI) -> AsyncIterator[None]:
-    if not PRODUCTION_MODE:
+    if not PRODUCTION_MODE and not DATABASE_MIGRATED:
         async with engine.begin() as conn:
-            if DB_URL.endswith("app.db") or DATABASE_RESET:
-                await conn.run_sync(Base.metadata.drop_all)
-            if not DB_URL.startswith("postgresql") or DATABASE_RESET:
-                await conn.run_sync(Base.metadata.create_all)
+            await conn.run_sync(Base.metadata.drop_all)
+            await conn.run_sync(Base.metadata.create_all)
+
+    loop: AbstractEventLoop = get_running_loop()
+    connection = await connect_robust(MQ_URL, loop=loop)
+    await pochta_producer.connect(connection)
 
     yield
+    await connection.close()
 
 
 app = FastAPI(lifespan=lifespan)
@@ -42,7 +48,7 @@ app.add_middleware(
 
 # API
 app.include_router(reglog_rst.router, prefix="/api")
-app.include_router(users_rst.router, prefix="/api/users")
+app.include_router(current_user_rst.router, prefix="/api/users/current")
 app.include_router(sessions_rst.router, prefix="/api/sessions")
 
 # MUB
